+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - widgets
     - frame: add `allow` property
+    - matrix: make `quantity` a dynamic property to avoid losing widget state when it changes
 
 ## 1.29.1 / 1.29.2